画像が保存されている場合の処理追加#92

+++ src/app/Http/Controllers/UserController.php
@@ -51,11 +51,12 @@ class UserController extends Controller
 
     public function update(UserRequest $request, User $user)
     {
-
         $image = $request->file('image');
         if ($image !== null) {
             $filename = $image->getClientOriginalName();
-            Storage::disk('public')->delete(($user->image));
+            if ($user->image !== null) {
+                Storage::disk('public')->delete(($user->image));
+            }
             InterVentionImage::make($image)->resize(
                 64,
                 64,
